--- components/ARP826/CMakeLists.txt
@@ -1,16 +1,0 @@
-set(reqs
-  EFCP
-  Rmt
-  NetworkInterface
-  ShimIPCP
-  BufferManagement
-  configSensor
-  Common
-  Portability
-  RINA_API
-  IPCP)
-
-idf_component_register(
-  SRCS "ARP826.c"
-  INCLUDE_DIRS "include"
-  REQUIRES ${reqs})
