+++ components/IPCP/CMakeLists.txt
@@ -1,4 +1,4 @@
 idf_component_register(
-  SRCS "IpcManager.c" "normalIPCP.c" "IPCP.c" "common.c"
-  INCLUDE_DIRS "include"
-  REQUIRES configSensor NetworkInterface ShimIPCP BufferManagement ARP826 Rmt RINA_API EFCP Enrollment Ribd FlowAllocator)
+    SRCS "IpcManager.c" "normalIPCP.c" "IPCP.c" "common.c"
+    INCLUDE_DIRS "include"
+    REQUIRES configSensor NetworkInterface ShimIPCP BufferManagement Rmt RINA_API EFCP Enrollment Ribd FlowAllocator)
