+++ components/ShimIPCP/Arp826/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(reqs
+  Rina-manager
+  Portability
+)
+
+idf_component_register(
+  SRCS "Arp826.c"
+  INCLUDE_DIRS "include"
+  REQUIRES ${reqs})
